recuiter_project_app/test.py
import google.generativeai as genai
import os
from dotenv import load_dotenv
load_dotenv()
genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

model = genai.GenerativeModel("gemini-2.0-flash")
resp = model.generate_content("Say hello from Gemini 1.5!")
print(resp.text)
